+++ Tests/Integration/MSFT_xMountImage_VHD.Integration.Tests.ps1
@@ -1,7 +1,3 @@
-# In order to run these tests, Hyper-V must be installed on the testing computer.
-# If it is not installed these tests will not be run. This does prevent these tests
-# from being run on AppVeyor.
-
 $script:DSCModuleName      = 'xStorage'
 $script:DSCResourceName    = 'MSFT_xMountImage'
 
@@ -26,21 +22,15 @@ $TestEnvironment = Initialize-TestEnvironment `
 # Using try/finally to always cleanup even if something awful happens.
 try
 {
-    # Ensure that the tests can be performed on this computer
-    if (-not (Test-HyperVInstalled))
-    {
-        Return
-    }
-
     #region Integration Tests for VHD
     # Get a spare drive letter
     $LastDrive = ((Get-Volume).DriveLetter | Sort-Object | Select-Object -Last 1)
     $DriveLetter = [char](([int][char]$LastDrive)+1)
 
-    # Create a VHDx with a partition
+    # Create a VHD with a partition
     $VHDPath = Join-Path -Path $ENV:Temp `
-        -ChildPath 'TestDisk.vhdx'
-    $null = New-VHD -Path $VHDPath -SizeBytes 10GB -Dynamic
+        -ChildPath 'TestDisk.vhd'
+    $null = New-VDisk -Path $VHDPath -SizeInMB 1024
     $null = Mount-DiskImage -ImagePath $VHDPath
     $disk = Get-Disk | Where-Object -Property Location -eq -Value $VHDPath
     $null = $disk | Initialize-Disk -PartitionStyle GPT
@@ -87,7 +77,7 @@ try
                 }
                 $current.Imagepath        | Should Be $VHDPath
                 $current.DriveLetter      | Should Be $DriveLetter
-                $current.StorageType      | Should Be 'VHDX'
+                $current.StorageType      | Should Be 'VHD'
                 $current.Access           | Should Be 'ReadWrite'
                 $current.Ensure           | Should Be 'Present'
             }
@@ -126,7 +116,7 @@ try
         }
     }
 
-    # Delete the VHDx test file created
+    # Delete the VHD test file created
     Remove-Item -Path $VHDPath -Force
     #endregion Integration Tests for VHD
 }