+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - Fix error when used on Windows Server 2012 R2 - See [Issue 102](https://github.com/PowerShell/xStorage/issues/102).
 - Added the VS Code PowerShell extension formatting settings that cause PowerShell
   files to be formatted as per the DSC Resource kit style guidelines.
+- Removed requirement on Hyper-V PowerShell module to execute integration tests.
 
 ## 3.1.0.0